refactor: extract gateway admin governance
